backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
cohere==5.13.0
qdrant-client==1.12.2
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.0
markdown==3.7
beautifulsoup4==4.12.3
numpy==2.2.0
typing-extensions==4.12.2
python-multipart==0.0.20